Added reference to java version in annotation.cfg Modified application.py so that it also returns the system name Fixed scikit integration, so that it is compatible with older stored ranking classifiers Removed stderr appearance of a sentence string because it still break unicode Added to translate.py the parameter to write standard and verbose output into a translation file

+++ src/mt/hybrid.py
@@ -158,5 +158,5 @@ class SimpleWsdTriangleTranslator(Worker):
         #print "Rank: ", rank
         
         for rank_item, output in zip(rank, outputs_ordered):
-            if float(rank_item)==1:
+            if int(rank_item)==1:
                 return output, description
